move LRU cachae to a separate file and native::detail namespace
