+++ aten/src/ATen/cuda/detail/CUDAHooks.cpp
@@ -5,7 +5,7 @@
 #include <ATen/Error.h>
 #include <ATen/RegisterCUDA.h>
 #include <ATen/cuda/CUDAConfig.h>
-#include <ATen/native/cuda/SpectralOps.h>
+#include <ATen/native/cuda/CuFFTPlanCache.h>
 #include <ATen/cuda/PinnedMemoryAllocator.h>
 #include <ATen/detail/CUDAHooksInterface.h>
 
@@ -167,19 +167,19 @@ double CUDAHooks::batchnormMinEpsilonCuDNN() const {
 }
 
 int64_t CUDAHooks::cuFFTGetPlanCacheMaxSize() const {
-  return at::native::__cufft_get_plan_cache_max_size_impl();
+  return at::native::detail::cufft_get_plan_cache_max_size_impl();
 }
 
 void CUDAHooks::cuFFTSetPlanCacheMaxSize(int64_t max_size) const {
-  at::native::__cufft_set_plan_cache_max_size_impl(max_size);
+  at::native::detail::cufft_set_plan_cache_max_size_impl(max_size);
 }
 
 int64_t CUDAHooks::cuFFTGetPlanCacheSize() const {
-  return at::native::__cufft_get_plan_cache_size_impl();
+  return at::native::detail::cufft_get_plan_cache_size_impl();
 }
 
 void CUDAHooks::cuFFTClearPlanCache() const {
-  at::native::__cufft_clear_plan_cache_impl();
+  at::native::detail::cufft_clear_plan_cache_impl();
 }
 
 int CUDAHooks::getNumGPUs() const {
